requirements.txt
djangorestframework>=3.9.0
django>=2.1.4
flake8==3.6.0
pyflakes==2.0.0
setuptools==39.0.1
django-model-utils==3.1.2
flake8-per-file-ignores==0.8.1
django-nose==1.4.6
coverage==4.5.3
django-cors-headers==3.0.2
psycopg2-binary==2.8.3
uWSGI==2.0.18
cryptography==2.8
celery==4.3.0
redis==3.3.11
crontab==0.22.6
requests==2.22.0